+++ CHANGELOG.md
@@ -29,3 +29,4 @@
 * (ERICK) Added lists for Take and Teach.
 * (ERICK) Created `addStudentToModule()` method.
 * (ERICK) Created `addStaffToModule()` method.
+* (ERICK) Created `removeStudentFromModule()` method.